chore: rename Tab
